requirements.txt
streamlit==1.56.0
pandas==3.0.2
plotly==6.7.0
numpy==2.4.4
scikit-learn==1.7.2
joblib==1.5.2
google-genai==1.47.0
